experiments/gemini-vision-eval/requirements.txt
google-genai>=0.6.0
pillow>=10.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
